+++ requirements-base.txt
@@ -1,0 +1,11 @@
+openai==1.64.0
+google-generativeai==0.8.0
+openpyxl==3.1.5
+pandas==2.2.3
+protobuf==5.29.3
+python-dotenv==1.0.1
+tqdm==4.67.1
+urllib3==2.3.0
+wcwidth==0.2.13
+vllm==0.9.0
+transformers==4.52.4
