+++ requirements-vllm.txt
@@ -1,0 +1,4 @@
+
+vllm==0.9.0
+
+
